class name has been changed
